+++ src/frontend/app/shared-tdoc/shared-tdoc.module.ts
@@ -48,7 +48,6 @@ import {GpxEditAreaComponent} from './components/gpx-editarea/gpx-editarea.compo
 import {TourDocReplaceFormComponent} from './components/tdoc-replaceform/tdoc-replaceform.component';
 import {TourDocAssignFormComponent} from './components/tdoc-assignform/tdoc-assignform.component';
 import {TourDocKeywordTagFormComponent} from './components/tdoc-keywordtagform/tdoc-keywordtagform.component';
-import {CommonDocObjectNavigationComponent} from './components/cdoc-object-navigation/cdoc-object-navigation.component';
 import {TourDocSimpleSearchNavigationComponent} from './components/tdoc-simple-search-navigation/tdoc-simple-search-navigation.component';
 
 @NgModule({
@@ -89,7 +88,6 @@ import {TourDocSimpleSearchNavigationComponent} from './components/tdoc-simple-s
         TourDocReplaceFormComponent,
         TourDocAssignFormComponent,
         TourDocKeywordTagFormComponent,
-        CommonDocObjectNavigationComponent,
         TourDocSimpleSearchNavigationComponent
     ],
     imports: [
@@ -148,7 +146,6 @@ import {TourDocSimpleSearchNavigationComponent} from './components/tdoc-simple-s
         TourDocReplaceFormComponent,
         TourDocAssignFormComponent,
         TourDocKeywordTagFormComponent,
-        CommonDocObjectNavigationComponent,
         TourDocSimpleSearchNavigationComponent
     ]
 })